Execute compiled values without formal round trips (#371)
* Execute compiled values without formal round trips

* Keep generated execution on compiled values

* Carry typed compiler failures to diagnostics

* Collapse duplicate algorithm validation

# Conflicts:
#	Sources/SwiftTLA/SpecParser.swift

+++ Tests/SwiftTLATests/CompiledLowererTraversalTests.swift
@@ -87,7 +87,7 @@ struct CompiledLowererTraversalTests {
         let operation = operations.first,
         arguments.count == 1,
         let argument = arguments.first,
-        case .value(.value(.int(3))) = argument,
+        case .value(.value(.integer(3))) = argument,
         case .boundValue(let bodyBinder) = operation.body,
         let parameter = operation.parameters.first
         else {
